[DEV-004] chnage float to int

+++ app.py
@@ -9,7 +9,7 @@ def add(a,b) :
 
 if __name__ == "__main__":
     print("Hello world !")
-    n =  float(   input("inter a number")   ) 
+    n =  int(   input("inter a number")   ) 
 
     n2 = calculate_square(n)
     print(f"result is {n2}")
